prevent warning and error when deleting consultation slots, fixes #932
Closes #932

+++ lib/models/ConsultationSlot.php
@@ -51,7 +51,7 @@ class ConsultationSlot extends SimpleORMap
         };
         $config['registered_callbacks']['after_delete'][] = function ($slot) {
             $block = $slot->block;
-            if (count($block->slots) === 0) {
+            if ($block && count($block->slots) === 0) {
                 $block->delete();
             }
         };
